seed/requirements.txt
psycopg2-binary>=2.9.9
faker==19.13.0
numpy>=1.26.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
